Renaming tfcontroller to tensorflow
