+++ pkg/controller.v2/tensorflow/informer.go
@@ -1,4 +1,4 @@
-package tfcontroller
+package tensorflow
 
 import (
 	"fmt"
