+++ pkg/controller.v2/tensorflow/status.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // Package controller provides a Kubernetes controller for a TFJob resource.
-package tfcontroller
+package tensorflow
 
 import (
 	"fmt"
@@ -53,10 +53,8 @@ func updateStatusSingle(tfjob *tfv1alpha2.TFJob, rtype tfv1alpha2.TFReplicaType,
 		tfjob.Status.StartTime = &now
 	}
 
-	// If the TFJob contains Chief or Master spec, then we will update the status
-	// according to the Chief/Master spec.
-	if ContainChieforMasterSpec(tfjob) {
-		if tfv1alpha2.IsChieforMaster(rtype) {
+	if ContainChiefSpec(tfjob) {
+		if rtype == tfv1alpha2.TFReplicaTypeChief {
 			if running > 0 {
 				msg := fmt.Sprintf("TFJob %s is running.", tfjob.Name)
 				err := updateTFJobConditions(tfjob, tfv1alpha2.TFJobRunning, tfJobRunningReason, msg)
