+++ pkg/controller.v2/tensorflow/util.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tfcontroller
+package tensorflow
 
 import (
 	"fmt"
@@ -40,11 +40,8 @@ func GetPortFromTFJob(tfJob *tfv1alpha2.TFJob, rtype tfv1alpha2.TFReplicaType) (
 	return -1, errPortNotFound
 }
 
-// ContainChieforMasterSpec returns true if the tfjob contains chief or master spec.
-func ContainChieforMasterSpec(tfJob *tfv1alpha2.TFJob) bool {
+func ContainChiefSpec(tfJob *tfv1alpha2.TFJob) bool {
 	if _, ok := tfJob.Spec.TFReplicaSpecs[tfv1alpha2.TFReplicaTypeChief]; ok {
-		return true
-	} else if _, ok := tfJob.Spec.TFReplicaSpecs[tfv1alpha2.TFReplicaTypeMaster]; ok {
 		return true
 	}
 	return false
